use the deployed treeview config for the sample
when I migrated the app to a hybrid solution, I modified the sample
to be what the config eventually should look like. but the code
is still using the old language.

P.S. the only difference between the new and the old so far is just
the support for multiple configuration (the part where the configs
are not directly added to the object and instead of using a key-value
pair and given a proper name).