Update dependencies (#222)
* ensure all included dependencies support react 19

to do this, I updated the dependencies that specifically had defined
react versions below 19. which were

- react-select
- react-window
- @mui/material
- @mui/x-tree-view

* update how styles are attached to SimpleTreeView

* use the deployed treeview config for the sample

when I migrated the app to a hybrid solution, I modified the sample
to be what the config eventually should look like. but the code
is still using the old language.

P.S. the only difference between the new and the old so far is just
the support for multiple configuration (the part where the configs
are not directly added to the object and instead of using a key-value
pair and given a proper name).